UP de: > * Diagrama de clases. > * Diagrama de secuencia. > * Diagrama de deployment. > * Correccion de ortografia y criterios de aceptacion de user stories. > * Correccion de ortografia de Retrospective. > * Ajuste en Test. > * Se agrego el tag v3.0

+++ Entrega nro 3/ImplementacionUserStories/Implementacion Travel Book/src/Program/TrasladoAereo.java
@@ -15,10 +15,10 @@ public class TrasladoAereo extends Traslado{
     private Usuario user;
     private Vector<TrasladoAereo> escalas = new Vector<TrasladoAereo>();
     
-    
-    protected static final Coordenada coordenada = new Coordenada ((float)41.2891,(float) 2.013);
-    protected static final Coordenada coordenadaLlegada = new Coordenada ((float) 12.31,(float) 345.65);
-    protected static final Coordenada coordenadaSalida = new Coordenada ((float) 876.31,(float) 453.65);
+    //Datos para simular FlightStats
+    protected static final Coordenada coordenada = new Coordenada (41.2891,2.013);
+    protected static final Coordenada coordenadaLlegada = new Coordenada (12.31,345.65);
+    protected static final Coordenada coordenadaSalida = new Coordenada (876.31,453.65);
     
     
     public TrasladoAereo(String nroVuelo, String compania, Calendar fechainicio, Usuario user, Aeropuerto aeroLlegadaTEST,
